% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel_reporter.R
\name{call_excel}
\alias{call_excel}
\title{Provides detailed diagnostic tests including tests for stationarity, autocorrelation for each model}
\usage{
call_excel(
  model_list,
  output_file = NULL,
  output_dir = getwd(),
  report_title = "Corporate",
  ...
)
}
\arguments{
\item{output_file}{name of the output file without extension}

\item{output_dir}{Name of the output directory, default is current directory}

\item{report_title}{Name of the report}

\item{modelList}{model list in LHS~RHS form}
}
\value{
Results of various diagnostic tets of Stationarity, BP, DW, Tests of significance etc provided.
}
\description{
Provides detailed diagnostic tests including tests for stationarity, autocorrelation for each model
}
\examples{
models <-c("DR_logit_FD ~ rl_est_ad+uae_ann_imp_yoy+uae_rl_cons_3qma","DR_logit_FD ~ uae_ann_imp_yoy+rl_est_ad_log+uae_rl_cons_3qma","DR_logit_FD ~ uae_ann_imp_yoy+rl_est_ad_3qma+uae_rl_cons_3qma")
call_excel(models,output_file = "Report2019") # allModelObjects created by modelDeveloper()
}
