% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallRmd.R
\name{call_rmd}
\alias{call_rmd}
\title{Knit the rmarkdown into an HTML document}
\usage{
call_rmd(
  model_list,
  output_file = "report.html",
  output_dir = getwd(),
  report_title = "Corporate",
  report_type = "html"
)
}
\arguments{
\item{output_file}{name of the output file}

\item{output_dir}{Name of the output directory}

\item{report_title}{Name of the report}

\item{modelList}{model list in LHS~RHS form}
}
\description{
Knit the rmarkdown into an HTML document
}
\examples{
models<-c("DR~avg_oil_pri_barrel_lag_3+avg_oil_pri_barrel_lag_2","DR~avg_oil_pri_barrel_lag_2","DR~Non_oil_ECI_yoy_ch_6QMA+avg_oil_pri_barrel_6QMA_lag_1")
validationSampler(macrodata,1:29,30:33,1:48)
call_rmd(models,report_title = "Retail")
}
