% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHACCorrectedStandardErrors.R
\name{checkHACCorrectedPValues}
\alias{checkHACCorrectedPValues}
\title{getFormulaofAllSignificantModels}
\usage{
checkHACCorrectedPValues(modelsList, trainingData, pvalueThreshold = 0.05)
}
\arguments{
\item{modelsList}{List of strings containing all formulae or model names}

\item{trainingData}{ususally train_df - this is used to build the models, ie the train data}

\item{pvalueThreshold}{the significance level usually 0.05 or 0.1}
}
\value{
Returns a list of all HAC adjusted p values
}
\description{
getFormulaofAllSignificantModels
}
\examples{
checkHACCorrectedPValues(allModelEvaluated_f$model,train_df)
}
