% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLinearityRamseyTest.R
\name{checkLinearityRamseyTest}
\alias{checkLinearityRamseyTest}
\title{checkLinearityRamseyTest}
\usage{
checkLinearityRamseyTest(modelsList, trainingData, pvalueThreshold = 0.05)
}
\arguments{
\item{modelsList}{List of strings containing all formulae or model names}

\item{trainingData}{ususally train_df - this is used to build the models, ie the train data}

\item{pvalueThreshold}{the significance level usually 0.05 or 0.1}
}
\value{
Returns a list of all  adjusted p values from Ramsey test
}
\description{
checkLinearityRamseyTest
}
\examples{
checkHACCorrectedPValues(allModelEvaluated_f$model,train_df)
}
