% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicConfig.R
\name{configDynamic}
\alias{configDynamic}
\title{Identify the response variable, the laggedvariable name and number of lags to be used.}
\usage{
configDynamic(depVar, lagVar, noofLags, trainIndex)
}
\arguments{
\item{depVar}{Dependent variable}

\item{lagVar}{Independent variable}

\item{noofLags}{No of lags}

\item{trainIndex}{index till which actual data is present}
}
\description{
Identify the response variable, the laggedvariable name and number of lags to be used.
}
