% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel_reporter.R
\name{dfSummaryWriter}
\alias{dfSummaryWriter}
\title{Creates summary sheet}
\usage{
dfSummaryWriter(file.name, sheetName = "Navigation", dataFrame)
}
\arguments{
\item{file.name}{Name of the excel file}

\item{sheetName}{name of the sheet}

\item{dataFrame}{Name of the dataframe to be written to the sheet}
}
\description{
Called in a loop depending on number of models
}
