% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel_reporter.R
\name{dfWriter}
\alias{dfWriter}
\title{Creates and formats sheets in Excel using openxslx. Called internally by call_excel function}
\usage{
dfWriter(file.name, sheetName, dataFrame)
}
\arguments{
\item{file.name}{Name of the excel file}

\item{sheetName}{name of the sheet}

\item{dataFrame}{Name of the dataframe to be written to the sheet}
}
\description{
Called in a loop depending on number of models
}
