% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicConfig.R
\name{dynamicCheck}
\alias{dynamicCheck}
\title{Check if dynamic model or not}
\usage{
dynamicCheck(selectedModel)
}
\arguments{
\item{depVar}{Dependent variable}

\item{lagVar}{Independent variable}

\item{noofLags}{No of lags}

\item{trainIndex}{index till which actual data is present}
}
\description{
Check if dynamic model or not
}
