% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicPredict.R
\name{dynamicForecast}
\alias{dynamicForecast}
\title{Dynamic forecasting function}
\usage{
dynamicForecast(
  selectedModel,
  selectedModelObject,
  laggedRelation,
  trainIndex = nrow(train_df)
)
}
\arguments{
\item{selectedModel}{Model defined by means of LHS and RHS(to be passed as a list)}

\item{selectedModelObject}{Default value of FALSE, single model to be examined.}

\item{laggedRelation}{Identify the response variable, the laggedvariable name and number of lags to be used. this has to be colon ':' separated and quoted-  for example "DR_logit_FD :DR_logit_FD_lag1:1" shows response variable is DR_logit_FD, the lagged independent variable is DR_logit_FD_lag1 and 1 indicates the number of lags to consider}

\item{trainIndex}{Index of rows for which all independent variables are present including the lagged one, typically traindata or training plus test data. Default is 1:nrow(train_df)}
}
\value{
dynamically forecast response variables
}
\description{
Dynamic forecasting function
}
\examples{
model <-"DR_logit_FD ~ DR_logit_FD_lag1 +uae_ann_imp_yoy+rl_est_ad_log+uae_rl_cons_3qma"
laggedRelation <- "DR_logit_FD :DR_logit_FD_lag1:1"
reporter(model,laggedRelation="DR_logit_FD :DR_logit_FD_lag1:1",trainIndex=1:30)

}
