% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatModelEval.R
\name{formatModelEvalResults}
\alias{formatModelEvalResults}
\title{Format model results of model eval}
\usage{
formatModelEvalResults(allModelEvaluated)
}
\arguments{
\item{allModelEvaluated}{model list with diagnostic tests}
}
\value{
condensedModelEval Condensed model eval
}
\description{
Format model results of model eval
}
\examples{
formatModelEvalResults(allModelEvaluated)
}
