% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormulaOfAllSignificantModels.R
\name{getFormulaofAllSignificantModels}
\alias{getFormulaofAllSignificantModels}
\title{getFormulaofAllSignificantModels}
\usage{
getFormulaofAllSignificantModels(allFormulae, train_df)
}
\arguments{
\item{allFormulae}{List of strings containing all formulae or model names}

\item{train_df}{Train df - this is used to build the models, ie the train data}
}
\value{
Returns a list of all formulae of models that are significant, the models themselves arent passed as this can be a very long list depending on the number of variables
}
\description{
getFormulaofAllSignificantModels
}
\examples{
getFormulaofAllSignificantModels(-c("z_index ~ cp_general_government_debt_to_gdp_ratio_fd",,train_df)
}
