% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormulaofAllSignificantModelsTurbo.R
\name{getFormulaofAllSignificantModelsTurbo}
\alias{getFormulaofAllSignificantModelsTurbo}
\title{getFormulaofAllSignificantModelsTurbo}
\usage{
getFormulaofAllSignificantModelsTurbo(
  allFormulae,
  train_df,
  pValueThreshold = 0.1,
  ncores = 8
)
}
\arguments{
\item{allFormulae}{List of strings containing all formulae or model names}

\item{train_df}{Train df - this is used to build the models, ie the train data}
}
\value{
Returns a list of all formulae of models that are significant, the models themselves arent passed as this can be a very long list depending on the number of variables
}
\description{
getFormulaofAllSignificantModelsTurbo
}
\examples{
getFormulaofAllSignificantModels(-c("z_index ~ cp_general_government_debt_to_gdp_ratio_fd",train_df)
}
