% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedVariableCombinations.R
\name{getUnRelatedVariableCombinations}
\alias{getUnRelatedVariableCombinations}
\title{Get Unrelated variables combinations, to avoid using variable combinations in the RHS of the formula that are transformations of the same variable}
\usage{
getUnRelatedVariableCombinations(
  LHS_vars,
  baseVariables,
  train_df,
  numberOfVariables,
  strictMax = TRUE
)
}
\arguments{
\item{LHS_vars}{Left hand side variable name, the response variable as a vector eg c("DR")}

\item{baseVariables}{Names of the untransformed variables that you want to use in the RHS, so if there is oil_price and oil_price_ln, provide just oil_price, it will identify that oil_price_ln is a transformation using "like" operator}

\item{train_df}{Train df - this is used to get the list of all macrovariables that have a likeness to each variable in baseVariables}
}
\value{
Returns a list of all formulae that do not have related variables in the same formula
}
\description{
Get Unrelated variables combinations, to avoid using variable combinations in the RHS of the formula that are transformations of the same variable
}
\examples{
getUnRelatedVariableCombinations(c("DR"),c("oil_price",'gdp_ratio'),train_df)
}
