% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDeveloperTurbo.R
\name{modelDeveloperTurbo}
\alias{modelDeveloperTurbo}
\title{Generates all models using combination of dependent and independent variables provided}
\usage{
modelDeveloperTurbo(
  LHS_var,
  RHS_vars,
  multiple = TRUE,
  no_of_vars,
  modelsNamesList = c(),
  trainData = train_df,
  pvalueThreshold = 0.05,
  progressStep = 10000
)
}
\arguments{
\item{RHS_vars}{Right hand side of the regression. A vector of dependent variables,  macroeconomic variable names for instance. These should be available in the data provided to  validationSampler.}

\item{no_of_vars}{Maximum Number of dependent variables to be used in the model building.}

\item{modelsNamesList}{List of Models to be built, if this is not NA, then the previous arguments are ignored}

\item{trainData}{Training dataframe on which model is to be built}

\item{pvalueThreshold}{the Pvalue threshold to be used for instance 0.05}

\item{progressStep}{THe number of models creation completion, after which a progress statement is logged to the console. If 1000, afetr every 1000 models, the user is given a message}

\item{LHS_vars}{LHS of the regression function of the for.}
}
\value{
All model formulae that pass significance in the form Y ~ X1 + X2
}
\description{
Generates all models using combination of dependent and independent variables provided
}
