% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelDiagnoser.R
\name{modelDiagnostics}
\alias{modelDiagnostics}
\title{Provides detailed diagnostic tests including tests for stationarity, autocorrelation for each model}
\usage{
modelDiagnostics(allModelObjects, testData = test_df,
  direction_config = macrometa)
}
\arguments{
\item{allModelObjects}{model Objects as generated by the function modelDeveloper}

\item{testData}{Test data to be provided as data.frame. If blank, it will look for test_df created by ValidationSampler in the Global environment.}

\item{direction_config}{To capture the expected sign of the coefficients of regression. This is a two column dataframe or datatable. Column names must be mandatorily 'Variable' and 'Type'. Variable contains the names of independent variables. Type denotes the expected relationship with response variable.It can take on either 1 or -1. If 1, indicates positive correlation, and hence expected signs of the coefficient is positive. If -1, indicates negative correlation, and hence expected signs of coefficients for those variables is negative}
}
\value{
Results of various diagnostic tets of Stationarity, BP, DW, Tests of significance etc provided.
}
\description{
Provides detailed diagnostic tests including tests for stationarity, autocorrelation for each model
}
\examples{
modelDiagnostics(allModelObjects) # allModelObjects created by modelDeveloper()
}
