% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluator.R
\name{modelEvaluator}
\alias{modelEvaluator}
\title{Evaluates and classifies models into pass or fail}
\usage{
modelEvaluator(allModelsDiagnostics, Heteroskedasticity_threshold = 0.1,
  normality_threshold = 0.1, pvalue_threshold = 0.05,
  Autocorrelation_threshold = 0.1, Stationarity_threshold = 0.1)
}
\arguments{
\item{Heteroskedasticity_threshold}{Threshold for pvalue for BP test}

\item{normality_threshold}{Threshold for pvalue for Shapiro test}

\item{pvalue_threshold}{Threshold for tests of significance}

\item{Autocorrelation_threshold}{Threshold for DurbinWatson test}

\item{Stationarity_threshold}{thresholds for ADF tests}
}
\value{
Consolidated model results with tests classified as PASS or fail
}
\description{
Different from modelDiagnoser in that the diagnoer only computed the diagnostics, it doesnt classify it
}
\examples{
modelEvaluator(allModelObjects) # allModelObjects created by modelDeveloper()
}
