% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFilter.R
\name{modelFilter}
\alias{modelFilter}
\title{Filter models based on adjusted Rsquared before running diagnostic tests. This is done merely to reduce the time taken by modelDiagnostic function}
\usage{
modelFilter(allModels, adj.r.squaredThreshold = 0.5)
}
\arguments{
\item{allModels}{allmodel objects as created by ModelDeveloper}

\item{adj.r.squaredThreshold}{Adjusted Rsquared threshold, models with less this threshold will be dropped from testing for other diagnostics}
}
\value{
Returns a list of selected model objects. These are not just names of models.
}
\description{
Filter models based on adjusted Rsquared before running diagnostic tests. This is done merely to reduce the time taken by modelDiagnostic function
}
\examples{
selectedModelObjs <- modelFilter(allModels,adj.r.squaredThreshold =0.5)
}
