% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPredictor.R
\name{ModelSensitiser}
\alias{ModelSensitiser}
\title{Sensitivity analyser}
\usage{
ModelSensitiser(
  selectedModel,
  selectedModelObject,
  predicted_df,
  sensitivity,
  rhs,
  forecast_df,
  ...
)
}
\arguments{
\item{selectedModel}{model name as a string as generated by the function modelDeveloper. String can be stored in a vector and used as well}

\item{selectedModelObject}{the results as generated by modelEvaluator function}

\item{predicted_df}{the results as generated from model predictor in baseline}

\item{sensitivity}{the sensitivity inputs from user, if not provided would default to plus minus 5%}

\item{rhs}{the vector of macrovariables that are to be sensitised}
}
\value{
Returns a list of dataframe with predicted values from different sensitivites of input
}
\description{
Sensitivity analyser
}
\examples{
ModelSensitiser(selectedModel,allModelEvaluated)
}
