% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configReport.R
\name{reporter}
\alias{reporter}
\title{Generates report details for each model defined by LHS and RHS}
\usage{
reporter(
  model,
  report_type = "html",
  multiple = FALSE,
  unTransformFunction = NA,
  ...
)
}
\arguments{
\item{model}{Model defined by means of LHS and RHS(to be passed as a list)}

\item{multiple}{Default value of FALSE, single model to be examined.}
}
\value{
List containing several report details. This is a named list
}
\description{
Generates report details for each model defined by LHS and RHS
}
\examples{
model_1<- list(LHS="DR", RHS=c("avg_oil_pri_barrel_lag_3","avg_oil_pri_barrel_lag_2"))
reporter(model_1)
}
