% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter_scenario_analyser.R
\name{reporter_scenario_analyser}
\alias{reporter_scenario_analyser}
\title{Generates report details for each model defined by LHS and RHS}
\usage{
reporter_scenario_analyser(reporter_output, ...)
}
\arguments{
\item{reporter_output}{output from reporter function ie. excel_details containing various diagnostic tests, scenario comparison}
}
\value{
results_of_overall_scenario dataframe containing results of diagnostic tests in a summary form
}
\description{
Generates report details for each model defined by LHS and RHS
}
\examples{
reporter_scenario_analyser(reporter_output)
}
