% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPredictor.R
\name{selectedModelCharter}
\alias{selectedModelCharter}
\title{Chart the model selected}
\usage{
selectedModelCharter(selectedModel, selectedModelObject, allModelEvaluated,
  ...)
}
\arguments{
\item{selectedModel}{model name as a string as generated by the function modelDeveloper. String can be stored in a vector and used as well}

\item{allModelEvaluated}{the results as generated by modelEvaluator function}
}
\value{
Returns ggplot
}
\description{
Chart the model selected
}
\examples{
selectedModelForecaster(selectedModel,allModelEvaluated)
}
