% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPredictor.R
\name{selectedModelForecaster}
\alias{selectedModelForecaster}
\title{Forecast values using chosen model}
\usage{
selectedModelForecaster(selectedModel, allModelEvaluated)
}
\arguments{
\item{selectedModel}{model name as a string as generated by the function modelDeveloper. String can be stored in a vector and used as well}

\item{allModelEvaluated}{the results as generated by modelEvaluator function}
}
\value{
Returns a dataframe with predicted values
}
\description{
Forecast values using chosen model
}
\examples{
selectedModelForecaster(selectedModel,allModelEvaluated)
}
