% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPredictor.R
\name{selectedModelScenariosCharter}
\alias{selectedModelScenariosCharter}
\title{Chart the scenarios predictions}
\usage{
selectedModelScenariosCharter(scenariosList, baseline_predictions, ...)
}
\arguments{
\item{scenariosList}{list containing all scenarios with scenario names and predictions}
}
\value{
Returns ggplot
}
\description{
Chart the scenarios predictions
}
\examples{
selectedModelScenariosCharter(scenariosList,baseline_predictions)
}
