% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPredictor.R
\name{selectedModelScenariosMEVCharter}
\alias{selectedModelScenariosMEVCharter}
\title{Chart the scenarios MEVs}
\usage{
selectedModelScenariosMEVCharter(
  scenariosList,
  baseline_predictions,
  mevs,
  ...
)
}
\arguments{
\item{scenariosList}{list containing all scenarios with scenario names and predictions}
}
\value{
Returns ggplot
}
\description{
Chart the scenarios MEVs
}
\examples{
selectedModelScenariosCharter(scenariosList,baseline_predictions)
}
