% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel_reporter.R
\name{sheetCreator}
\alias{sheetCreator}
\title{Creates and formats sheets in Excel using openxslx. Called internally by call_excel function}
\usage{
sheetCreator(model, file.name, sheetName, ...)
}
\arguments{
\item{model}{name of the model}

\item{file.name}{Name of the excel file}
}
\description{
Called in a loop depending on number of models
}
