% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformRuleAdd}
\alias{transformRuleAdd}
\title{Identify the response variable, and associated transformation rule}
\usage{
transformRuleAdd(depVar, transformRule)
}
\arguments{
\item{depVar}{Dependent variable}

\item{transformRule}{transformation rule}

\item{noofLags}{No of lags}

\item{trainIndex}{index till which actual data is present}
}
\description{
Identify the response variable, and associated transformation rule
}
