% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customTransform.R
\name{UnTransformFunctionWrapper}
\alias{UnTransformFunctionWrapper}
\title{Custom UnTransformFunction function}
\usage{
UnTransformFunctionWrapper(unransformFunction, ...)
}
\arguments{
\item{...}{other paramaters as necessary for the custom transform function}

\item{transformFunction}{Function that is to be called to transform}
}
\value{
Vector of converted values
}
\description{
Custom UnTransformFunction function
}
\examples{
model_1<- customTransform(zScoreCompute,...)
}
