% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customTransform.R
\name{untransformZscore}
\alias{untransformZscore}
\title{Zscore untransform function}
\usage{
untransformZscore(valuesToBeTransformed, rho, originalUntransformedValues)
}
\arguments{
\item{zScoreCompute}{Function that is to be called to transform to zscores from PDs}

\item{originalUntransformedPDs}{orignal pd values that were transformed to zscores, this is needed to get the mean of zscores which is used for untranssforming. for example in z_index ~ mev1+mev2,these are the values(actuals usually) used to obtain the z_index}

\item{zScoresTobeUntransformed}{Values to be untransformed to PDs, usually this is the output of regression for example in z_index ~ mev1+mev2,  this is the predicted z-index}
}
\value{
List of untransformed vales, usually represent the PDs
}
\description{
Zscore untransform function
}
\examples{
model_1<- untransformZscore(c(0.1,0.2),c(0.1,0.2),0.023)
}
