% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationSampler.R
\name{validationSampler}
\alias{validationSampler}
\title{Splits into training, test and forecast datasets}
\usage{
validationSampler(macroecodata, rn_train, rn_test, rn_forecast)
}
\arguments{
\item{macroecodata}{A Dataset with macroeconomic variables and values and Default rates.}

\item{rn_train}{Rowindex or rownumbers of macroecodata as vector that forms training dataset.}

\item{rn_test}{Rowindex or rownumbers of macroecodata as vector that forms testing dataset.}

\item{rn_forecast}{Rowindex or rownumbers of macroecodata as vector that forms training dataset.}
}
\value{
Training, test and forecast datasets in global Environment
}
\description{
Splits into training, test and forecast datasets
}
\examples{
validationSampler(macroecodata,1:29,30:33,1:48)
}
