% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableSelector.R
\name{variableSelector}
\alias{variableSelector}
\title{Filter variable list by running single variable regressions and selecting models based on R-squared higher than threshold (default threshold is 5%). This
runs the modelDeveloper function with no_of_vars  =  1}
\usage{
variableSelector(LHS_vars, RHS_vars, trainData = train_df,
  R2Threshold = 0.1)
}
\arguments{
\item{LHS_vars}{LHS of the regression function of the form Y ~ X1 + X2. This is the vector of response variable. Example default rates, or PIT PDs for default rate forecasting}

\item{RHS_vars}{Right hand side of the regression equationof the form Y ~ X1 + X2. A vector of dependent variables,  macroeconomic variable names for instance. These should be available in the data provided to  validationSampler.}

\item{trainData}{Training dataframe on which model is to be built}

\item{R2Threshold}{Threshold for Rsquared -default at 0.1}
}
\value{
Returns a vector with list of selected Variables
}
\description{
Filter variable list by running single variable regressions and selecting models based on R-squared higher than threshold (default threshold is 5%). This
runs the modelDeveloper function with no_of_vars  =  1
}
\examples{
variableSelector(LHS_vars = c("DR_logit_FD"),RHS_vars = vars,trainData =train_df)
}
