% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/createComboAndgetFormulaofAllSignificantModelsTurboFastLM.R
\name{createComboAndgetFormulaofAllSignificantModelsTurboFastLM}
\alias{createComboAndgetFormulaofAllSignificantModelsTurboFastLM}
\title{createComboAndgetFormulaofAllSignificantModelsTurboFastLM}
\usage{
createComboAndgetFormulaofAllSignificantModelsTurboFastLM(
  baseVars,
  numberOfVariables,
  yIndex,
  train_df,
  pValueThreshold = 0.1,
  ncores = 4,
  saveToDisk = TRUE
)
}
\arguments{
\item{numberOfVariables}{Number of variables to be used in each formula. See strictMax for 'upto' vs 'exact' match}

\item{yIndex}{Index of the response variable column in train_df}

\item{train_df}{Train df - this is used to get the list of all macrovariables that have a likeness to each variable in baseVariables}

\item{baseVariables}{Names of the untransformed variables that you want to use in the RHS, so if there is oil_price and oil_price_ln, provide just oil_price, it will identify that oil_price_ln is a transformation using "like" operator}
}
\value{
Returns a list of all formulae of models that are significant, the models themselves arent passed as this can be a very long list depending on the number of variables
}
\description{
createComboAndgetFormulaofAllSignificantModelsTurboFastLM
}
\examples{
createComboAndgetFormulaofAllSignificantModelsTurboFastLM(-c("z_index ~ cp_general_government_debt_to_gdp_ratio_fd",2,train_df)
}
