% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedVariableCombinations.R
\name{getUnRelatedVariableCombinations}
\alias{getUnRelatedVariableCombinations}
\title{Get Unrelated variables combinations, to avoid using variable combinations in the RHS of the formula that are transformations of the same variable}
\usage{
getUnRelatedVariableCombinations(
  LHS_vars,
  baseVariables,
  train_df,
  numberOfVariables,
  strictMax = TRUE,
  patternType = "startsWith"
)
}
\arguments{
\item{LHS_vars}{Left hand side variable name, the response variable as a vector eg c("DR")}

\item{baseVariables}{Names of the untransformed variables that you want to use in the RHS, so if there is oil_price and oil_price_ln, provide just oil_price, it will identify that oil_price_ln is a transformation using "like" operator}

\item{train_df}{Train df - this is used to get the list of all macrovariables that have a likeness to each variable in baseVariables}

\item{numberOfVariables}{Number of variables to be used in each formula. See strictMax for 'upto' vs 'exact' match}

\item{strictMax}{Default as TRUE, if false, it looks for upto numberoFvariables, if TRUE, it looks at exact number of variables in the combinations it generates}

\item{patternType}{default as "startsWith". If any other, then it uses %like%}
}
\value{
Returns a list of all formulae that do not have related variables in the same formula
}
\description{
Get Unrelated variables combinations, to avoid using variable combinations in the RHS of the formula that are transformations of the same variable
}
\examples{
getUnRelatedVariableCombinations(c("DR"),c("oil_price",'gdp_ratio'),train_df)
}
