% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationSampler.R
\name{modelDeveloper}
\alias{modelDeveloper}
\title{Generates all models using combination of dependent and independent variables provided}
\usage{
modelDeveloper(LHS_vars, RHS_vars, no_of_vars, trainData = train_df)
}
\arguments{
\item{LHS_vars}{LHS of the regression function of the form Y ~ X1 + X2. This is the vector of response variable. Example default rates, or PIT PDs for default rate forecasting}

\item{RHS_vars}{Right hand side of the regression equationof the form Y ~ X1 + X2. A vector of dependent variables,  macroeconomic variable names for instance. These should be available in the data provided to  validationSampler.}

\item{no_of_vars}{Maximum Number of dependent variables to be used in the model building.}

\item{trainData}{Maximum Number of dependent variables to be used in the model building.}
}
\value{
All model objects.Note that this is not a dataframe but S3 objects. If you want to explore particular models, then one can use the subset operator $
}
\description{
Generates all models using combination of dependent and independent variables provided
}
